scripts/pubsub-tobigquery-function/requirements.txt
google-cloud-bigquery==2.3.1
